+++ lib/capybara/apparition/page/keyboard.rb
@@ -66,6 +66,10 @@ module Capybara::Apparition
 
   private
 
+    def insert_emoji(str)
+      @page.command('Input.insertText', text: str)
+    end
+
     def type_with_modifiers(keys, delay:)
       old_pressed_keys, @pressed_keys = @pressed_keys, {}
 
@@ -74,9 +78,17 @@ module Capybara::Apparition
         when Array
           type_with_modifiers(sequence, delay: delay)
         when String
-          sequence.each_char do |char|
-            press char
-            sleep delay
+          clusters = sequence.grapheme_clusters.chunk { |gc| gc.match?(/\p{Emoji Presentation}/) }
+          clusters.each do |emoji, chars|
+            if emoji
+              insert_emoji(chars.join)
+              sleep delay
+            else
+              chars.each do |char|
+                press char
+                sleep delay
+              end
+            end
           end
         else
           press sequence
@@ -91,7 +103,7 @@ module Capybara::Apparition
     end
 
     def release_pressed_keys
-      @pressed_keys.values.each { |desc| up(desc) }
+      @pressed_keys.each_value { |desc| up(desc) }
     end
 
     def key_description(key)
@@ -330,6 +342,7 @@ module Capybara::Apparition
       'z': { 'keyCode': 90, 'code': 'KeyZ', 'shiftKey': 'Z', 'key': 'z' },
       'meta': { 'keyCode': 91, 'key': 'Meta', 'code': 'MetaLeft' },
       'command': { 'keyCode': 91, 'key': 'Meta', 'code': 'MetaLeft' },
+      'cmd': { 'keyCode': 91, 'key': 'Meta', 'code': 'MetaLeft' },
       '*': { 'keyCode': 106, 'key': '*', 'code': 'NumpadMultiply', 'location': 3 },
       '+': { 'keyCode': 107, 'key': '+', 'code': 'NumpadAdd', 'location': 3 },
       '-': { 'keyCode': 109, 'key': '-', 'code': 'NumpadSubtract', 'location': 3 },