+++ lib/capybara/apparition/page/mouse.rb
@@ -9,13 +9,14 @@ module Capybara::Apparition
       @current_buttons = BUTTONS[:none]
     end
 
-    def click_at(x:, y:, button: 'left', count: 1, modifiers: [])
+    def click_at(x:, y:, button: 'left', count: 1, delay: 0, modifiers: [])
       move_to x: x, y: y
       count.times do |num|
         @keyboard.with_keys(modifiers) do
           mouse_params = { x: x, y: y, button: button, count: num + 1 }
-          down mouse_params
-          up mouse_params
+          down(**mouse_params)
+          sleep(delay || 0)
+          up(**mouse_params)
         end
       end
       self
@@ -29,7 +30,7 @@ module Capybara::Apparition
 
     def down(button: 'left', **options)
       options = @current_pos.merge(button: button).merge(options)
-      mouse_event('mousePressed', options)
+      mouse_event('mousePressed', **options)
       @current_buttons |= BUTTONS[button.to_sym]
       self
     end
@@ -37,7 +38,7 @@ module Capybara::Apparition
     def up(button: 'left', **options)
       options = @current_pos.merge(button: button).merge(options)
       @current_buttons &= ~BUTTONS[button.to_sym]
-      mouse_event('mouseReleased', options)
+      mouse_event('mouseReleased', **options)
       self
     end