+++ spec/integration/driver_spec.rb
@@ -4,6 +4,7 @@ require 'spec_helper'
 require 'pdf/reader'
 require 'chunky_png'
 require 'fastimage'
+require 'logger'
 require 'os'
 
 module Capybara::Apparition
@@ -40,6 +41,27 @@ module Capybara::Apparition
       end
     end
 
+    context 'output redirection with Ruby logger' do
+      let(:io) { StringIO.new }
+      let(:logger) { Logger.new(io) }
+      let(:session) { Capybara::Session.new(:apparition_with_logger, TestApp) }
+
+      before do
+        Capybara.register_driver :apparition_with_logger do |app|
+          Capybara::Apparition::Driver.new(app, browser_logger: logger)
+        end
+      end
+
+      after do
+        session.driver.quit
+      end
+
+      it 'supports capturing console.log' do
+        session.visit('/apparition/console_log')
+        expect(io.string).to include('Hello world')
+      end
+    end
+
     it 'quits silently before visit call' do
       driver = Capybara::Apparition::Driver.new(nil)
       expect { driver.quit }.not_to raise_error
@@ -73,22 +95,20 @@ module Capybara::Apparition
       end
 
       it 'allows custom maximization size' do
-        begin
-          Capybara.register_driver :apparition_with_custom_screen_size do |app|
-            Capybara::Apparition::Driver.new(
-              app,
-              logger: TestSessions.logger,
-              screen_size: [800, 600]
-            )
-          end
-          session = Capybara::Session.new(:apparition_with_custom_screen_size, TestApp)
-          session.visit(session_url('/'))
-          session.current_window.resize_to(400, 400)
-          session.current_window.maximize
-          expect(session.current_window.size).to eq([800, 600])
-        ensure
-          session&.driver&.quit
+        Capybara.register_driver :apparition_with_custom_screen_size do |app|
+          Capybara::Apparition::Driver.new(
+            app,
+            logger: TestSessions.logger,
+            screen_size: [800, 600]
+          )
         end
+        session = Capybara::Session.new(:apparition_with_custom_screen_size, TestApp)
+        session.visit(session_url('/'))
+        session.current_window.resize_to(400, 400)
+        session.current_window.maximize
+        expect(session.current_window.size).to eq([800, 600])
+      ensure
+        session&.driver&.quit
       end
     end
 
@@ -103,22 +123,20 @@ module Capybara::Apparition
     end
 
     it 'supports specifying viewport size with an option' do
-      begin
-        Capybara.register_driver :apparition_with_custom_window_size do |app|
-          Capybara::Apparition::Driver.new(
-            app,
-            logger: TestSessions.logger,
-            window_size: [800, 600]
-          )
-        end
-        driver = Capybara::Session.new(:apparition_with_custom_window_size, TestApp).driver
-        driver.visit(session_url('/'))
-        expect(
-          driver.evaluate_script('[window.innerWidth, window.innerHeight]')
-        ).to eq([800, 600])
-      ensure
-        driver&.quit
+      Capybara.register_driver :apparition_with_custom_window_size do |app|
+        Capybara::Apparition::Driver.new(
+          app,
+          logger: TestSessions.logger,
+          window_size: [800, 600]
+        )
       end
+      driver = Capybara::Session.new(:apparition_with_custom_window_size, TestApp).driver
+      driver.visit(session_url('/'))
+      expect(
+        driver.evaluate_script('[window.innerWidth, window.innerHeight]')
+      ).to eq([800, 600])
+    ensure
+      driver&.quit
     end
 
     shared_examples 'render screen' do
@@ -380,7 +398,7 @@ module Capybara::Apparition
         expect(@driver.evaluate_script('window.navigator.userAgent')).to eq('foo')
       end
 
-      it 'sets headers for all HTTP requests', :headers2 do
+      it 'sets headers for all HTTP requests' do
         @driver.headers = { 'X-Omg' => 'wat' }
         @session.visit '/'
         sleep 1 # ensure page loaded
@@ -396,8 +414,7 @@ module Capybara::Apparition
             document.body.innerHTML = t;
           });
         JS
-        sleep 2 # time for XHR request to run and update body
-        expect(@driver.body).to include('X_OMG: wat')
+        expect(@session).to have_content('X_OMG: wat')
       end
 
       it 'adds new headers' do
@@ -410,17 +427,16 @@ module Capybara::Apparition
         expect(@driver.body).to include('APPENDED: true')
       end
 
-      it 'sets headers on the initial request', :headers3 do
+      it 'sets headers on the initial request' do
         skip 'Need to figure out the timing on this' if ENV['CI']
         @driver.headers = { 'PermanentA' => 'a' }
         @driver.add_headers('PermanentB' => 'b')
         @driver.add_header('Referer', 'http://google.com', permanent: false)
         @driver.add_header('TempA', 'a', permanent: false)
-
         @session.visit('/apparition/headers_with_ajax')
 
-        initial_request = @session.find(:css, '#initial_request').text
-        ajax_request = @session.find(:css, '#ajax_request').text
+        initial_request = @session.find(:css, '#initial_request', text: /.+/, wait: 3).text
+        ajax_request = @session.find(:css, '#ajax_request', text: /.+/, wait: 3).text
 
         expect(initial_request).to include('PERMANENTA: a')
         expect(initial_request).to include('PERMANENTB: b')
@@ -581,17 +597,15 @@ module Capybara::Apparition
       end
 
       it 'errors when extension is unavailable' do
-        begin
-          @failing_driver = Capybara::Apparition::Driver.new(
-            @session.app,
-            logger: TestSessions.logger,
-            inspector: !ENV['DEBUG'].nil?,
-            extensions: %W[#{File.expand_path '../support/non_existent.js', __dir__}]
-          )
-          expect { @failing_driver.visit '/' }.to raise_error(Capybara::Apparition::BrowserError, /Unable to load extension: .*non_existent\.js/)
-        ensure
-          @failing_driver.quit
-        end
+        @failing_driver = Capybara::Apparition::Driver.new(
+          @session.app,
+          logger: TestSessions.logger,
+          inspector: !ENV['DEBUG'].nil?,
+          extensions: %W[#{File.expand_path '../support/non_existent.js', __dir__}]
+        )
+        expect { @failing_driver.visit '/' }.to raise_error(Capybara::Apparition::BrowserError, /Unable to load extension: .*non_existent\.js/)
+      ensure
+        @failing_driver.quit
       end
     end
 
@@ -626,28 +640,24 @@ module Capybara::Apparition
       end
 
       it 'does not propagate a Javascript error to ruby if error raising disabled' do
-        begin
-          driver = Capybara::Apparition::Driver.new(@session.app, js_errors: false, logger: TestSessions.logger)
-          driver.visit session_url('/apparition/js_error')
-          driver.execute_script 'setTimeout(function() { omg }, 0)'
-          sleep 0.1
-          expect(driver.body).to include('hello')
-        ensure
-          driver&.quit
-        end
+        driver = Capybara::Apparition::Driver.new(@session.app, js_errors: false, logger: TestSessions.logger)
+        driver.visit session_url('/apparition/js_error')
+        driver.execute_script 'setTimeout(function() { omg }, 0)'
+        sleep 0.1
+        expect(driver.body).to include('hello')
+      ensure
+        driver&.quit
       end
 
       it 'does not propagate a Javascript error to ruby if error raising disabled and client restarted' do
-        begin
-          driver = Capybara::Apparition::Driver.new(@session.app, js_errors: false, logger: TestSessions.logger)
-          driver.restart
-          driver.visit session_url('/apparition/js_error')
-          driver.execute_script 'setTimeout(function() { omg }, 0)'
-          sleep 0.1
-          expect(driver.body).to include('hello')
-        ensure
-          driver&.quit
-        end
+        driver = Capybara::Apparition::Driver.new(@session.app, js_errors: false, logger: TestSessions.logger)
+        driver.restart
+        driver.visit session_url('/apparition/js_error')
+        driver.execute_script 'setTimeout(function() { omg }, 0)'
+        sleep 0.1
+        expect(driver.body).to include('hello')
+      ensure
+        driver&.quit
       end
     end
 
@@ -777,7 +787,7 @@ module Capybara::Apparition
         @session.visit('/apparition/with_js')
         expect(@driver.network_traffic.length).to eq(4).or eq(5) # 4 plus potential favicon
         @driver.clear_network_traffic
-        expect(@driver.network_traffic.reject { |t| /favicon.ico$/.match? t.url }.length).to eq(0)
+        expect(@driver.network_traffic.count { |t| !(/favicon.ico$/.match? t.url) }).to eq(0)
       end
 
       it 'blocked requests get cleared along with network traffic' do
@@ -949,33 +959,29 @@ module Capybara::Apparition
     end
 
     it 'allows the driver to have a fixed port' do
-      begin
-        driver = Capybara::Apparition::Driver.new(@driver.app, port: 12345)
-        driver.visit session_url('/')
+      driver = Capybara::Apparition::Driver.new(@driver.app, port: 12345)
+      driver.visit session_url('/')
 
-        expect { TCPServer.new('127.0.0.1', 12345) }.to raise_error(Errno::EADDRINUSE)
-      ensure
-        driver.quit
-      end
+      expect { TCPServer.new('127.0.0.1', 12345) }.to raise_error(Errno::EADDRINUSE)
+    ensure
+      driver.quit
     end
 
     it 'allows the driver to have a custom host' do
-      begin
-        # Use custom host "pointing" to localhost, specified by APPARITION_TEST_HOST env var.
-        # Use /etc/hosts or iptables for this: https://superuser.com/questions/516208/how-to-change-ip-address-to-point-to-localhost
-        # A custom host and corresponding env var for Travis is specified in .travis.yml
-        # If var is unspecified, skip test
-        host = ENV['APPARITION_TEST_HOST']
+      # Use custom host "pointing" to localhost, specified by APPARITION_TEST_HOST env var.
+      # Use /etc/hosts or iptables for this: https://superuser.com/questions/516208/how-to-change-ip-address-to-point-to-localhost
+      # A custom host and corresponding env var for Travis is specified in .travis.yml
+      # If var is unspecified, skip test
+      host = ENV['APPARITION_TEST_HOST']
 
-        skip 'APPARITION_TEST_HOST not set' if host.nil?
+      skip 'APPARITION_TEST_HOST not set' if host.nil?
 
-        driver = Capybara::Apparition::Driver.new(@driver.app, host: host, port: 12_345)
-        driver.visit session_url('/')
+      driver = Capybara::Apparition::Driver.new(@driver.app, host: host, port: 12_345)
+      driver.visit session_url('/')
 
-        expect { TCPServer.new(host, 12_345) }.to raise_error(Errno::EADDRINUSE)
-      ensure
-        driver&.quit
-      end
+      expect { TCPServer.new(host, 12_345) }.to raise_error(Errno::EADDRINUSE)
+    ensure
+      driver&.quit
     end
 
     it 'lists the open windows' do
@@ -1415,6 +1421,14 @@ module Capybara::Apparition
         expect(input.value).to eq('apparition')
       end
 
+      it 'sends emojis' do
+        input = @session.find(:css, '#empty_input')
+
+        input.native.send_keys('a😀cd😴 🛌🏽🇵🇹 e🤾🏽‍♀️f')
+
+        expect(input.value).to eq('a😀cd😴 🛌🏽🇵🇹 e🤾🏽‍♀️f')
+      end
+
       it 'has an alias' do
         input = @session.find(:css, '#empty_input')
 
@@ -1528,6 +1542,31 @@ module Capybara::Apparition
             'c' => @session.driver.find_css('#empty_textarea, #filled_textarea')
           }
         )
+      end
+    end
+
+    context 'with remote browser' do
+      let(:options) { { 'remote-debugging-address' => '127.0.0.1', 'remote-debugging-port' => rand(9000..10000) } }
+
+      before do
+        @remote_browser = Capybara::Apparition::Browser::Launcher::Local.start(headless: true, browser_options: options)
+        @remote_browser.ws_url
+
+        @extended_driver = Capybara::Apparition::Driver.new(
+          @session.app,
+          remote: true,
+          browser_options: options
+        )
+      end
+
+      after do
+        @extended_driver.quit
+        @remote_browser.stop
+      end
+
+      it 'supports rendering the page' do
+        @extended_driver.visit session_url('/')
+        expect(@extended_driver.body).to include('Hello world!')
       end
     end
   end