+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+# Version 0.6.0
+Release date: 2020-06-21
+
+### Added
+
+* Support Ruby loggers in the :logger and :browser_logger option [Daniel Orner]
+* Support "Remote" Chrome browser [Volodymyr Mykhailyk]
+* Support rapid set functionality
+* Support click delay
+* Experimental: Support sending emoji to text inputs
+
+### Fixed
+
+* User Fetch CDP domain for network interception
+
+### Changed
+
+* Only support Ruby 2.5+
+
+# Version 0.5.0
+Release date: 2020-01-26
+
+### Added
+
+* Support :drop_modifiers option in `#drag_to`
+* Support setting range input
+
+### Fixed
+
+* Ruby 2.7 keyword arguments warnings
+* Error in visibility JS atom [Stefan Wienert]
+* Issue with request headers [dabrowt1]
+
 # Version 0.4.0
 Release date: 2019-07-15